refactor: move next cookies module to separate directory
